organize the ssr server files
