+++ src/server/package.json
@@ -5,8 +5,8 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "DEBUG=app:* nodemon src/server/index",
-    "start": "NODE_ENV=production nodemon src/server/index",
+    "dev": "DEBUG=app:* nodemon index",
+    "start": "NODE_ENV=production node index",
     "test:ci": "echo \"Error: no test specified\" && exit 0"
   },
   "repository": {
