+++ src/server/utils/auth/strategies/oauth.js
@@ -3,7 +3,7 @@ const axios = require("axios");
 const boom = require("@hapi/boom");
 const { OAuth2Strategy } = require("passport-oauth");
 
-const { config } = require("../../../../config");
+const { config } = require("../../../config");
 
 const GOOGLE_AUTHORIZATION_URL = "https://accounts.google.com/o/oauth2/v2/auth";
 const GOOGLE_TOKEN_URL = "https://www.googleapis.com/oauth2/v4/token";
